+++ src/__experimental__/components/label/index.d.ts
@@ -1,1 +1,1 @@
-export { default } from './label';
+export { default } from "./label";