+++ src/components/accordion/accordion.spec.js
@@ -16,7 +16,6 @@ import {
   StyledAccordionContainer,
   StyledAccordionSubTitle,
   StyledAccordionTitleContainer,
-  StyledAccordionTitle,
   StyledAccordionIcon,
   StyledAccordionContent,
   StyledAccordionContentContainer,
@@ -525,51 +524,6 @@ describe("Accordion", () => {
       );
     });
   });
-
-  describe("style overrides", () => {
-    const randomStyleObject = {
-      backgroundColor: "red",
-      display: "flex",
-      fontSize: "200px",
-    };
-    beforeEach(() => {
-      render({
-        styleOverride: {
-          root: randomStyleObject,
-          headerArea: randomStyleObject,
-          icon: randomStyleObject,
-          header: randomStyleObject,
-          content: randomStyleObject,
-        },
-      });
-    });
-
-    it("renders root element with properly assigned styles", () => {
-      assertStyleMatch(
-        randomStyleObject,
-        wrapper.find(StyledAccordionContainer)
-      );
-    });
-
-    it("renders header area element with properly assigned styles", () => {
-      assertStyleMatch(
-        randomStyleObject,
-        wrapper.find(StyledAccordionTitleContainer)
-      );
-    });
-
-    it("renders icon element with properly assigned styles", () => {
-      assertStyleMatch(randomStyleObject, wrapper.find(StyledAccordionIcon));
-    });
-
-    it("renders header element with properly assigned styles", () => {
-      assertStyleMatch(randomStyleObject, wrapper.find(StyledAccordionTitle));
-    });
-
-    it("renders content element with properly assigned styles", () => {
-      assertStyleMatch(randomStyleObject, wrapper.find(StyledAccordionContent));
-    });
-  });
 });
 
 describe("AccordionGroup", () => {
@@ -703,5 +657,17 @@ describe("AccordionGroup", () => {
       " type `Accordion`.\n    in AccordionGroup";
 
     expect(console.error).toHaveBeenCalledWith(expected); // eslint-disable-line no-console
+  });
+
+  it("accepts empty children", () => {
+    expect(() => {
+      mount(
+        <AccordionGroup>
+          {null}
+          {false}
+          {undefined}
+        </AccordionGroup>
+      );
+    }).not.toThrow();
   });
 });