+++ src/components/dialog-full-screen/dialog-full-screen.spec.js
@@ -13,6 +13,7 @@ import { assertStyleMatch } from "../../__spec_helper__/test-utils";
 import IconButton from "../icon-button";
 import StyledIconButton from "../icon-button/icon-button.style";
 import { StyledHeader, StyledHeading } from "../heading/heading.style";
+import Help from "../help";
 
 jest.mock("../../utils/helpers/guid");
 
@@ -38,13 +39,44 @@ describe("DialogFullScreen", () => {
     );
   });
 
+  describe("contentRef", () => {
+    it("the content ref should be forwarded", () => {
+      let mockRef;
+
+      const WrapperComponent = () => {
+        mockRef = useRef();
+
+        return (
+          <DialogFullScreen
+            onCancel={onCancel}
+            className="foo"
+            open
+            title="my title"
+            subtitle="my subtitle"
+            contentRef={mockRef}
+          >
+            <Button>Button</Button>
+            <Button>Button</Button>
+          </DialogFullScreen>
+        );
+      };
+
+      wrapper = mount(<WrapperComponent />);
+
+      expect(mockRef.current).toBe(wrapper.find(StyledContent).getDOMNode());
+    });
+  });
+
   describe("autoFocus", () => {
+    jest.useFakeTimers();
     it("should focus the first element by default", () => {
       mount(
         <DialogFullScreen open>
           <input type="text" />
         </DialogFullScreen>
       );
+
+      jest.runAllTimers();
 
       const firstFocusableElement = document.querySelector("input");
       expect(document.activeElement).toBe(firstFocusableElement);
@@ -57,6 +89,8 @@ describe("DialogFullScreen", () => {
         </DialogFullScreen>
       );
 
+      jest.runAllTimers();
+
       const firstFocusableElement = document.querySelector("input");
       expect(document.activeElement).not.toBe(firstFocusableElement);
     });
@@ -64,6 +98,7 @@ describe("DialogFullScreen", () => {
 
   describe("focusFirstElement", () => {
     it("should focus on the element passes as focusFirstElement prop", () => {
+      jest.useFakeTimers();
       const Component = () => {
         const secondInputRef = useRef(null);
         return (
@@ -74,6 +109,8 @@ describe("DialogFullScreen", () => {
         );
       };
       mount(<Component />);
+
+      jest.runAllTimers();
 
       const secondFocusableElement = document.querySelectorAll("input")[1];
       expect(document.activeElement).toEqual(secondFocusableElement);
@@ -202,6 +239,20 @@ describe("DialogFullScreen", () => {
         const heading = fullScreenHeading.find(Heading);
 
         expect(heading.props().title).toEqual("my custom heading");
+      });
+    });
+
+    describe("when prop help is passed", () => {
+      it("should render Help component", () => {
+        wrapper = mount(
+          <DialogFullScreen
+            open
+            title="This is test title"
+            help="this is help text"
+          />
+        );
+
+        expect(wrapper.find(Help).exists()).toBe(true);
       });
     });
   });