+++ src/components/fieldset/fieldset.component.js
@@ -1,34 +1,26 @@
 import React from "react";
 import PropTypes from "prop-types";
+import styledSystemPropTypes from "@styled-system/prop-types";
+
 import { validProps } from "../../utils/ether/ether";
 import tagComponent from "../../utils/helpers/tags/tags";
+import { filterStyledSystemMarginProps } from "../../style/utils";
 import {
   FieldsetStyle,
   LegendContainerStyle,
   FieldsetContentStyle,
 } from "./fieldset.style";
-import Logger from "../../utils/logger/logger";
 
-let deprecatedWarnTriggered = false;
+const marginPropTypes = filterStyledSystemMarginProps(
+  styledSystemPropTypes.space
+);
 
 const Fieldset = (props) => {
-  if (!deprecatedWarnTriggered) {
-    deprecatedWarnTriggered = true;
-    // eslint-disable-next-line max-len
-    Logger.deprecate(
-      "`styleOverride` that is used in the `Fieldset` component is deprecated and will soon be removed."
-    );
-  }
-
   const legend = () => {
     if (!props.legend) return null;
 
     return (
-      <LegendContainerStyle
-        inline={props.inline}
-        data-component="legend-style"
-        styleOverride={props.styleOverride.legend}
-      >
+      <LegendContainerStyle inline={props.inline} data-component="legend-style">
         <legend data-element="legend">{props.legend}</legend>
       </LegendContainerStyle>
     );
@@ -43,7 +35,8 @@ const Fieldset = (props) => {
     <FieldsetStyle
       {...tagComponent("fieldset", props)}
       {...safeProps}
-      styleOverride={props.styleOverride.root}
+      m={0}
+      {...filterStyledSystemMarginProps(props)}
     >
       <FieldsetContentStyle
         data-component="fieldset-style"
@@ -57,22 +50,18 @@ const Fieldset = (props) => {
 };
 
 Fieldset.propTypes = {
+  /** Filtered styled system margin props */
+  ...marginPropTypes,
   /** Child elements */
   children: PropTypes.node,
   /** The text for the fieldsets legend element. */
   legend: PropTypes.string,
   /** When true, legend is placed in line with the children */
   inline: PropTypes.bool,
-  /** Allows to override existing component styles */
-  styleOverride: PropTypes.shape({
-    root: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-    legend: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
-  }),
 };
 
 Fieldset.defaultProps = {
   inline: false,
-  styleOverride: {},
 };
 
 export default Fieldset;