+++ src/components/menu/menu-divider/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from './menu-divider';
-export * from './menu-divider';
+export { default } from "./menu-divider";
+export * from "./menu-divider";