+++ src/components/text-editor/__internal__/editor-link/index.d.ts
@@ -1,2 +1,2 @@
-export { default } from './editor-link';
-export * from './editor-link';
+export { default } from "./editor-link";
+export * from "./editor-link";